Refactoring and stubbing for custom data parsing
