+++ src/main.rs
@@ -29,7 +29,7 @@ async fn main() {
 
     let data = Data::new(
         initialize_database().await,
-        Arc::new(game_data::parser::initialize_data().await),
+        Arc::new(game_data::parser::multi_source_parser::parse_data().await),
     )
     .await;
